chore(rust): clippy clean-up — 20 warnings → 0 (excl. workspace-profile noise)
`cargo clippy --release --lib` flagged 20 warnings across the engine
crate. Fixed all 19 in-code ones; the remaining one
("profiles for the non root package will be ignored") is a Cargo.toml
workspace-structure quirk, not source code, and survives any change.

### Auto-applied via `cargo clippy --fix --allow-dirty` (14 fixes)

- `rand::Rng` unused import in `dna.rs:19` and `fault.rs:11` → removed
- `lgssm.rs:97,130` "taken reference of right operand" → deref simplified
- `optimizer.rs:190,221` reference-immediately-dereferenced → simplified
- `optimizer.rs:296` collapsible `if` → merged
- `photonic.rs:61` unnecessary cast `usize → usize` → removed
- 5 trivial reference / deref simplifications across lib.rs

### Hand-fixed (5 fixes)

- `photonic.rs:22 MZI_INSERTION_LOSS_DB` unused → `#[allow(dead_code)]`
  with comment "reserved for future MZI loss accounting" (the constant
  is part of the public photonic loss budget, removing it changes the
  documented loss model)
- `photonic.rs:61` unnecessary outer parens around manhattan distance
  expression → removed
- `dna.rs:29 DEFAULT_TOEHOLD` unused → `#[allow(dead_code)]` with
  comment "reserved for strand-displacement toehold heuristic"
- `optimizer.rs:84` `accuracy.max(0.1).min(1.0)` clamp-pattern →
  `accuracy.clamp(0.1, 1.0)`
- `quantum.rs:163 + lib.rs:4798` `(Vec<(usize, f64)>, Vec<((usize,
  usize), f64)>)` type complexity → `#[allow(clippy::type_complexity)]`
  with comment "tuple shape mirrors Python's QUBO format" (changing
  the shape would break the Python QUBO API, type-aliasing the tuple
  loses the structural meaning at the API boundary)

### Verification

- `cargo clippy --release --lib`: 0 in-source warnings (only the
  workspace-profile noise remains, which is non-actionable)
- `cargo test --release --lib partition::tests`: 4/4 pass
- `cargo test --release --lib fault::tests`: all unit tests pass
- 218 chiplet pytest cases pass after rebuilding the maturin wheel

+++ engine/src/photonic.rs
@@ -19,6 +19,7 @@ use rayon::prelude::*;
 // ── Constants ────────────────────────────────────────────────────────
 
 const CROSSING_LOSS_DB: f64 = 0.08;
+#[allow(dead_code)] // reserved for future MZI loss accounting
 const MZI_INSERTION_LOSS_DB: f64 = 0.5;
 
 // ── Waveguide routing ────────────────────────────────────────────────
@@ -58,8 +59,8 @@ pub fn route_waveguides(
 
             let (ri, ci) = (i / grid_size, i % grid_size);
             let (rj, cj) = (j / grid_size, j % grid_size);
-            let manhattan = ((ri as isize - rj as isize).unsigned_abs()
-                + (ci as isize - cj as isize).unsigned_abs()) as usize;
+            let manhattan = (ri as isize - rj as isize).unsigned_abs()
+                + (ci as isize - cj as isize).unsigned_abs();
 
             let length_um = manhattan as f64 * pitch_um;
             let mut loss = length_um * 1e-4 * loss_db_per_cm;
